requirements.txt
python-vlc==3.0.7110
python-mpv==0.5.2
requests[socks]==2.24.0
grequests==0.6.0
pytest==5.3.5
CJKwrap==2.2
lxml==4.6.2
beautifulsoup4==4.7.1
coverage==5.0.3
codacy-coverage==1.3.11
flake8==3.7.9